Seperation of code to 'lib' & 'apiserver' folders
Modified the code generation folder structure.

All API server specific code is by default generated to apiserver
folder. If SERVER_TYPE env is set, it will be genreated to the folder
specified by SERVER_TYPE env. apiserver will contain the following code:
1. api/openapi.yaml
2. controllers/
    a. <module>.ts
    b. default.ts
    c. index.ts
    d. controllers.ts
3. services/
    a. <module>/apis.ts
    b. index.ts
    c. service.ts
4. index.ts -> server index file
5. server.ts -> Express server configuration file.

All the library code is stored in 'lib' folder. This folder contains:
1. docs
    a. model.png
    b. model.puml
    // Data base access code is stored in lib, so docs for model are also stored in lib.

2. drivers
    a. api-error.ts
    b. logger.ts
    c. morgan.ts
3. models
    a. <model>.ts
    b. <schema>.ts
    c. enums.ts
    d. index.ts
    e. plugins/
        i. index.ts
        ii. paginate.ts
        iii. to-json.ts
4. config.ts

The rest of the code is generated in the src folder.
1. nodemon.json
2. package.json
3. tsconfig.json